backend/requirements.txt
Flask
openai
spacy
python-docx
PyMuPDF
Flask-Login
Flask-Bcrypt
pymongo
python-dotenv
flask-cors
